Moved files arround, cleaned up, no changed behavior
